+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.2.1 [unreleased]
+## 2.2.1
 
 - Added `curve` option
 - Added `legend` option